requirements.txt
requests
aws-cdk-lib
constructs
aws-cdk.aws-apigatewayv2-alpha
aws_cdk.aws_apigatewayv2_integrations_alpha
